add variable scene functionality
A scenario is now defined with a class with a name and the array of scenes. Before this was only an array of scenes not in an class object.
Changed functions to allow any scenario to play via the scenario parameter.
Added dim div overlay that will dim when a scene is playing.
Removed play scene button from the scene window and added it to the wrapper/body. The button doesn't disappear when pressed, but summons the scene window above it to play the scene. NEEDS FURTHER WORK.