Change skip functionality and change button handling to be dynamic
Skip button now can only be used (before it couldn't be) while the scene is playing. Makes the scene counter go to the confession scene count and guides the scene there.

Added master button function that takes in an argument for the scenario number, assigned by the html side of the button. THIS MIGHT NEED TO CHANGE LATER?

Need to stop music when skip is engaged.
Need to check whether string is empty when clicking on the button that accepts user input.