add characters remaining counter
Added eventhandlers for inputting and deleting text inside the text area. A constant is defined early to be the max character length that the user can input, a number equal to this number minus the current characters they have in the text area to show how many characters are left.