add new data field, -animation- for Scene
Added an array of animation objects for a Scene. These animation objects contain an image, and animation shorthand values, i.e. animation duration, etc.

In the playScene function, a function animateBG loops through the array and makes all of the animations for that scene.

BUG FIX: Allow the input message to show up correctly when spoken. The onboundary event fires twice when at the beginning of the sentence, I have the part that handles the case in which it is speaking to ignore onboundary if the event.charIndex is the same as the previous fire.

Also fixed how it appears in the markup. It is no longer inside the scene-bg div, because that one expands and moves. It is a direct child of scene, and it will also not overflow.