Add code generation for scenario
Added simple encode and decode funcitons. Encode fires when the scene ends. This grabs the scene id, the scene's pitch and rate, as well as the input message, which is caeser ciphered by a key number. The code is then concatenated into one string. The key to the decipher is added to the end. The ciphering is to mainly not give a way the spoiler of the scenario too soon in the URL.

This code is given at the end of the scenario. A window pop up and the code is inside a input text field. CURRENTLY, THERE IS A BUG where the code might have some null values, when it should say '0'. I'll have to check on that.

To use the code. Paste it after the URL and '?'. After that click the ready button, which will run the decode function. Populate the scene variables and play the scene. Currently I can't get the scene to just play without user interaction. This might be insurmountable due to modern and chrome policies about autoplay.