progress restructuring
soundIterator is now iterateSounds, this, playScenario and sceneEnd are methods of Scenario. animateEntity is a method of SceneAnimation.

Got the program to play up to voiceinput box appearing again.

Need to figure out how to handle proposalSim as a whole. New idea is to just have one scenario inside proposaalSim, when it is closed, set it to null and when a button is pressed, set it to whatever.