add nextSounds array to ProposalSim
the nextSounds array are an array of variables that have setTimeout functions in them defined by whenever the iterateSounds runs. For every sound in the scene, it creates a variable to be pushed into the nextSounds array. In this variable/function, the sound is pushed onto the soundList property, and immediately plays. The separation of the soundList and nextSounds, ensures that the sounds can be cancelled while playing and while scheduled to be played.