separate scenario definitions from code
canvas.js is now propose-sim.js. scenarios.js takes a chunk out of propose-sim, the scenario definitions. Further restructured the base of the program. Now the close button works and the scene plays up to the very end scene.

Possible bug with the text message lingering if the window is closed as the words are spoken. Doesn't seem to have any other adverse effects though.