finish Animation Kitchen
The Kitchen Animation is finished. keyframes can have more transform effects, but they have to be on one line, before there were two declaration of transforms and the latter one overrode the other. This was fixed. The issue regarding remnants of the previous scenario ending scene showing up briefly has been fixed. In cleanUpVar, all divs with the class 'scene-bg' are removed. Also the end window and end divs are removed on cleanUpVar as well.