entity support
entities now can be fully customized. Specified the logo-end entity that will appear at the end of every scenario.

for the SceneAnimation object, new data fields put in to further specify the entity, including width, height, positionx and positiony, as well as the wild card 'special', which will take the argument and have it as the entity's additional class which will handle potential special effects.